Use PKCS12 as Keycloak keystore
The Java JKS keystore is deprecated, PKCS12 is the recommended
replacment. This patch removes the generation of the Java JKS
keystore. If a cert/key pair is provided they are copies onto the host
in the Keycloak configuration area. A PKCS12 file is then created in
that location. The cert/key files are no longer necessary on the host
after PKCS12 generation, as such they are removed. If a PKCS12 file is
provided it is copied to the configuration area on the host.

A new variable `keycloak_keystore_name` is introduced to control the
name of the PKCS12 keystore.

A keystore can contain any number of certs & keys, each entry can be
identified by name (e.g. alias). Best practice is to name all entries
and then reference the desired cert/key in the TLS configuration by
name, to this end the following new variable is introduced:

keycloak_tls_pkcs12_alias

In addition the CA certificate handling was removed, it was
unnecessary.

Because PKCS12 generation is now done on the target host and host
dependency on pyOpenSSL was added.

Signed-off-by: John Dennis <jdennis@redhat.com>